fix: add polyfill after chainable webpack has been executed (#34)
* fix: add polyfill after chainable webpack has been executed

* fix: use break to stop loop
