+++ packages/ice-scripts/lib/plugins/userConfig/configs/injectBabel.js
@@ -1,6 +1,4 @@
-const path = require('path');
-
-module.exports = ({ chainWebpack, context }, injectBabel) => {
+module.exports = ({ chainWebpack }, injectBabel) => {
   chainWebpack((config) => {
     if (injectBabel === 'runtime') {
       ['jsx', 'tsx'].forEach((rule) => {
@@ -26,21 +24,6 @@ module.exports = ({ chainWebpack, context }, injectBabel) => {
             return Object.assign(options, { plugins });
           });
       });
-    } else if (injectBabel === 'polyfill') {
-      const { rootDir } = context;
-      const entries = config.toConfig().entry;
-      const rule = config.module.rule('polyfill').test(/\.jsx?|\.tsx?$/);
-      Object.keys(entries).forEach((key) => {
-        // only include entry path
-        rule.include.add(path.resolve(rootDir, entries[key][0]));
-      });
-      rule.use('polyfill-loader').loader(require.resolve('../utils/polyfillLoader')).options({});
-
-      // add resolve modules for get core-js and regenerator-runtime
-      const modulePath = require.resolve('core-js');
-      const pathArr = modulePath.split('node_modules');
-      pathArr.pop(); // pop file path
-      config.resolve.modules.add(path.join(pathArr.join('node_modules'), 'node_modules'));
     }
   });
 };
